Float is not precise enough at large magnitudes for Fixed --we need double! Changed tests to reflect this

+++ tests/addition.cpp
@@ -27,10 +27,10 @@ TEST_CASE("Fixed += Fixed") {
     );
     Fixed foo(i);
     Fixed bar(j);
-    float foo_f = (float)foo, bar_f = (float)bar;
+    double foo_f = (double)foo, bar_f = (double)bar;
     foo += bar;
     // allowed to deviate up to the smallest step in the fixed-point representation
-    REQUIRE((float)foo == Approx(foo_f + bar_f).margin(1.0 / 4096));
+    REQUIRE((double)foo == Approx(foo_f + bar_f).margin(1.0 / 4096));
 }
 
 TEST_CASE("Fixed + Fixed") {
@@ -55,8 +55,8 @@ TEST_CASE("Fixed + Fixed") {
     );
     Fixed foo(i);
     Fixed bar(j);
-    float foo_f = (float)foo, bar_f = (float)bar;
+    double foo_f = (double)foo, bar_f = (double)bar;
     Fixed baz = foo + bar;
     // allowed to deviate up to the smallest step in the fixed-point representation
-    REQUIRE((float)baz == Approx(foo_f + bar_f).margin(1.0 / 4096));
+    REQUIRE((double)baz == Approx(foo_f + bar_f).margin(1.0 / 4096));
 }
